+++ Models/CalculatorData.cs
@@ -8,24 +8,24 @@ namespace intex2.Models
 {
     public class CalculatorData
     {
-        public int long_utm_x { get; set; }
-        public int lat_utm_y { get; set; }
-        public int hour { get; set; }
-        public int month { get; set; }
-        public int motorcycle_involved { get; set; }
-        public int pedestrian_involved { get; set; }
-        public int overturn_rollover { get; set; }
-        public int bicyclist_involved { get; set; }
-        public int teenage_driver_involved { get; set; }
+        public float long_utm_x { get; set; }
+        public float lat_utm_y { get; set; }
+        public float hour { get; set; }
+        public float month { get; set; }
+        public float motorcycle_involved { get; set; }
+        public float pedestrian_involved { get; set; }
+        public float overturn_rollover { get; set; }
+        public float bicyclist_involved { get; set; }
+        public float teenage_driver_involved { get; set; }
 
-        public Tensor<int> AsTensor()
+        public Tensor<float> AsTensor()
         {
-            int[] data = new int[]
+            float[] data = new float[]
             {
                 long_utm_x, lat_utm_y, hour, month, motorcycle_involved, pedestrian_involved, overturn_rollover, bicyclist_involved, teenage_driver_involved
             };
             int[] dimensions = new int[] { 1, 9 };
-            return new DenseTensor<int>(data, dimensions);
+            return new DenseTensor<float>(data, dimensions);
         }
     }
 }
